+++ src/GamePack/Trail.java
@@ -35,7 +35,7 @@ public class Trail extends GameObject{
         g2d.setComposite(makeTransparent(alpha));
 
         g.setColor(color);
-        g.fill3DRect(x,y,width,height,true);
+        g.fill3DRect((int)x,(int)y,width,height,true);
 
         g2d.setComposite(makeTransparent(1));
